move package version into a subpackage
